requirements.txt
google-genai
pillow
python-dotenv
opencv-python
qreader
fastapi
uvicorn
python-multipart
gTTS
